Refactor report atomic writes
